Add files via upload
Simple linear Regression task-
In this regression task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple linear regression task as it involves just two variables.
What Will be predicted score if a student studies for 9.25 hours in a day?